python_project/requirements.txt
azure_storage==0.37.0
Pillow==9.1.0
tkcalendar==1.6.1
